backend/pyproject.toml
[tool.poetry]
name = "linkpulse"
version = "0.3.0"
description = ""
authors = ["Xevion <xevion@xevion.dev>"]
license = "GNU GPL v3"
readme = "README.md"
package-mode = true

[tool.poetry.scripts]
app = "linkpulse"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.100"
python-dotenv = "^1.0.1"
peewee = "^3.17.7"
peewee-migrate = "^1.13.0"
types-peewee = "^3.17.7.20241017"
types-psycopg2 = "^2.9.21.20241019"
fastapi-cache2 = "^0.2.2"
questionary = "^2.0.1"
apscheduler = "^3.10.4"
human-readable = "^1.3.4"
psycopg2 = "^2.9.10"
structlog = "^24.4.0"
uvicorn = "^0.32.0"
asgi-correlation-id = "^4.3.4"
orjson = "^3.10.10"
hypercorn = "^0.17.3"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
pytest-xdist = "^3.6.1"
email-validator = "^2.2.0"
limits = "^3.13.0"
toml = "^0.10.2"
types-toml = "^0.10.8.20240310"


[tool.poetry.group.dev.dependencies]
memory-profiler = "^0.61.0"
bpython = "^0.24"
types-pytz = "^2024.2.0.20241003"
pytest = "^8.3.3"
httpx = "^0.27.2"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 110